requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pymongo==4.13.0
groq==0.3.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6 
pandas
langchain
langchain-community
langchain_groq
httpx
